+++ src/main/java/com/example/holaserver/Review/DTO/ReviewSaveRequestDto.java
@@ -1,5 +1,6 @@
-package com.example.holaserver.Review;
+package com.example.holaserver.Review.DTO;
 
+import com.example.holaserver.Review.Review;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
