--- src/main/java/com/example/holaserver/Review/ReviewTag.java
@@ -1,23 +1,0 @@
-package com.example.holaserver.Review;
-
-import com.example.holaserver.Common.BaseTimeEntity;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import java.sql.Timestamp;
-
-@Entity
-@NoArgsConstructor
-@Getter
-public class ReviewTag extends BaseTimeEntity {
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String tagText;
-
-    private Timestamp removedAt;
-}
