+++ src/main/java/com/example/holaserver/Review/ReviewTag/ReviewTag.java
@@ -1,26 +1,23 @@
-package com.example.holaserver.Review;
+package com.example.holaserver.Review.ReviewTag;
 
 import com.example.holaserver.Common.BaseTimeEntity;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.*;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 import java.sql.Timestamp;
 
 @Entity
-@Getter
 @NoArgsConstructor
-public class ReviewTagLog extends BaseTimeEntity {
+@Getter
+public class ReviewTag extends BaseTimeEntity {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Long reviewId;
-
-    private Long userId;
-
-    private Long storeId;
-
-    private Long reviewTagId;
+    private String tagText;
 
     private Timestamp removedAt;
 }
