+++ src/main/java/com/example/holaserver/Review/ReviewTagLog/ReviewTagLogRepository.java
@@ -1,5 +1,6 @@
-package com.example.holaserver.Review;
+package com.example.holaserver.Review.ReviewTagLog;
 
+import com.example.holaserver.Review.ReviewTagLog.ReviewTagLog;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
